Compare L2 addresses with their representations
Integrating different systems sometimes requires comparing
L2 addresses. That makes a problem when data to compare is
a complex data structure and the integration script should
be agnostic to its internals.

This patch allows to compare EUI objects for equality and
unequality with their valid representations and therefore
makes integration of several different systems easier.