Fix bad pop() assumption in test
Do not assume that IPSet.pop will always return the same item.  The
pop method is implemented via the popitem method on the underlying
_cidrs object, which is a dict.

dict.popitem is documented as:

Remove and return an arbitrary (key, value) pair from the dictionary.

So one cannot assert that the value popped from an IPSet will be a
specific item in the set.

Instead, assert that the item popped out of the set is present in the
other set.