Added __bytes__ implementation for intuitive usage
When constructing a bytes object from an IPAddress object in Python3
it seems intuitive for the result to be a byte string of the IP address
words.

e.g.

>>> bytes(IPAddress('1.1.1.1'))
b'\x01\x01\x01\x01'

Instead, __index__ is called and perhaps millions of bytes are allocated,
which can even crash unaware users. This can be surprising and unexpected.

>>> bytes(IPAddress('1.1.1.1'))
b'\x00\x00\x00\x00... # is this something we should expect?

In addition, having bytes(IPAddress(...)) yield a byte string is far
more readable, idiomatic, and is highly applicable to usage, such as
storing IPAddresses in databases, without using a one-shot socket and
struct import.

Implementing __bytes__ allows for high readability in this use case.