Use raw strings for escape characters used in regex
There were a couple instances of strings using the '\' character.
Starting with Python 3.6, normal strings that contain these types of
characters cause DeprecationWarnings to be raised, and presumably will
eventually stop allowing them completely.

This updates the two cases of using the '\' character to be raw strings
to allow it without extra special escaping of the characters.

Closes #184

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>