Replaced self-made machinery for retrieval of version from hardcoded value to `setuptools_scm`, which retrieves the version from git tags, and generates the file with harrdcoded version.
The benefit of this approach is following:

* we use a single source of truth, and it is a git tag, eliminating the possibility of inconsistency, when a git tag is set to one version, and the hardcoded value is another one.
* `setuptools_scm` for non-tag commits also stores part of commithash as a part of version. Then this part of commit hash can be retrieved, that provides additional information in the case of troubleshooting.

The benefit of using a reusable plugin to setuptools are

* we can then eliminate setup.py entirely (yes, I have kept it, but in fact it is completely not needed). setup.py is itself a security risk (yes, I'm a paranoid, and it is actually a good thing ;), and yes, this already have been exploited in the wil

d), especially when pip is run from root, it was one of the main reasons behind declarative metadata with rigid schema. By using a plugin we shift the trust from potentially malicious setup.py supplied by a package maintainer to a well-known plugin us

ed b
y lot of packages and checked by many people and supplied by a third party other than the maintainer of the package being installed.
* better code reuse