Implement IPAddress.to_canonical()
This is one of the pieces missing to have a good IPv4-mapped IPv6
address handling story.

Following Rust's IpAddr::to_canonical() design[1][2][3] which seems
sensible to me (except that ours is simpler as we don't have separate
types for IPv4 and IPv6 addresses – yet).

Also, relying on doctests[4] to test this feature – no need to make things
more difficult than they need to be.

[1] https://doc.rust-lang.org/std/net/enum.IpAddr.html#method.to_canonical
[2] https://doc.rust-lang.org/std/net/struct.Ipv6Addr.html#method.to_canonical
[3] https://github.com/rust-lang/rust/pull/87708
[4] 3290b50baa64 ("Make pytest run doctests (#280)")