Reorganize the documentation
I've been big fan of the Diataxis documentation framework, both as
a documentation reader and a writer.

Some of the content we have doesn't necessarily fit one of the modes
*perfectly* but that's fine, it's a start.

Notable I left the Introduction section outside of the modes (I intend
to remove it altogether) and the Authors and Contributors sections I'm
not sure what to do about.

I also changed one document's title so that it fits the format expected
by the How-to guides (How to ...).

[1] https://diataxis.fr/