Refactor IP category tests
There was quite a bit of repetition there and for the addresses we have
in the test code we were only verifying some of their is_*() values.

Considering I want to add some new is_*() methods I figured I'd refactor
the code first so that the test suite is more robust and it's easier to
discover invalid assumptions or other mistakes.

I got the refactoring idea after reading the Rust core library tests for
a similar feature[1].

The addresses we test are the same as before, just reordered.

[1] https://github.com/rust-lang/rust/blob/767453eb7ca188e991ac5568c17b984dd4893e77/library/core/tests/net/ip_addr.rs#L475C12-L475C12