Migrate from setup.{py,cfg} to pyproject.toml
Something I wanted to do for over a year now. setup.py is somewhat
outdated now and there's new tooling to build packages[1].

The package metadata in pyproject.toml is as close to what we had in
setup.py as the new format allows (I had to remove the platforms
property and the URL declaration is a bit different, maybe something
else I now forgot about).

Heavily inspired by two PRs made by KOLANICH and parts of this PR were
lifted from then, just with a different syntax and formatting.

[1] https://drivendata.co/blog/python-packaging-2023
[2] https://github.com/netaddr/netaddr/pull/227
[3] https://github.com/netaddr/netaddr/pull/230