Remove the IPAddress.is_private method
It was mixing different unrelated address ranges but they weren't even
all address ranges that could be considered "private" in some context.

All in all seemed like opportunity for confusion and mistakes.

There are better replacements for a good subset of that (mentioned in
the changelog).

In reality is_private existed in IPnetwork as well but

a) is_private() returning a bool on a network could never quite work
   correctly (there's no simple yes/no answer there)
b) It was not documented as supported