Fix running the test suite on musl systems
musl's inet_ntop doesn't format IPv4-compatible IPv6 addresses in any
special way and there's no dot notation in the output there.

We need to account for that.

In doctests this is achieved by changing the examples or, where not
applicable, disabling them.

In more standard test code we simply made a few tests conditional.

Resolves: https://github.com/netaddr/netaddr/issues/381